+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -67,6 +67,7 @@ public protocol HIGTheme: Sendable {
     var list: any HIGListTokens { get }
     var sidebar: any HIGSidebarTokens { get }
     var adminShell: any HIGAdminShellTokens { get }
+    var codeBlock: any HIGCodeBlockTokens { get }
     var photoPicker: any HIGPhotoPickerTokens { get }
     var photoEditor: any HIGPhotoEditorTokens { get }
     var longTextEditor: any HIGLongTextEditorTokens { get }